+++ src/allotropy/parsers/qiacuity_dpcr/qiacuity_dpcr_parser.py
@@ -16,8 +16,7 @@ class QiacuitydPCRParser(VendorParser):
 
     @property
     def release_state(self) -> ReleaseState:
-        # Waiting on more test data to validate before releasing
-        return ReleaseState.CANDIDATE_RELEASE
+        return ReleaseState.RECOMMENDED
 
     def to_allotrope(self, named_file_contents: NamedFileContents) -> Model:
         reader = QiacuitydPCRReader(named_file_contents.contents)