+++ src/allotropy/parsers/thermo_fisher_qubit4/thermo_fisher_qubit4_parser.py
@@ -50,7 +50,7 @@ class ThermoFisherQubit4Parser(VendorParser):
 
     @property
     def release_state(self) -> ReleaseState:
-        return ReleaseState.WORKING_DRAFT
+        return ReleaseState.RECOMMENDED
 
     def to_allotrope(self, named_file_contents: NamedFileContents) -> Model:
         """