+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Change Cedex HiRes to recommended release state
+- Change Qubit Flex to recommended release state
+- Change Qubit 4 ASM to recommended release state
+- Change MabTech Apex to recommended release state
+- Change Qiacuity to recommended release state
+
 ### Fixed
 
 ### Changed